Start implementing the Translate Component,
Start implementing the Translate Component, which will use our already existing Dropdown Component (making a few minor adjustments) and also use a new Convert Component we will focus on a little bit later. Then start working on the Convert Component. Pass in the ‘language’ and ‘text’ props. Take advantage of the useEffect hook to run code when ‘language’ or ‘text’ changes. Make a test call to the Google API based on the ‘language’ and ‘text’ props.