Set up a Navigation Header from scratch to navigate between our widgets.
Clicking on a link in the Header will not induce a hard-reload (full-page reload) of the react app. We will implement a special ‘Link’ inside of our Header Component to make sure page refreshes do not happen. Link will also detect changes in the window pathname, and also allow users to ‘Ctrl-Click’ it to open it in a new tab.